Update Dockerfile
In the refactored Dockerfile:

Comments are added to provide a brief description of each step.

The instructions are indented for better readability.

The COPY command is kept as is, but if there are other files or directories needed for the application, you can add them to the WORKDIR.

The CMD instruction specifies the command to run the Spring Boot application as an executable JAR file.

This Dockerfile sets up a Docker image for running a Spring Boot application. It uses OpenJDK 11 as the base image, copies the JAR file into the container, exposes port 8080 to allow external access, and sets the command to run the Spring Boot application using the "java -jar" command.